Moved application registration logic to ApplicationsManager.
Moved application registration logic for service, client and aks applications to ApplicationsManager. Made MicrosoftGraphServiceClient into a generic client class without application specific logic.
