+++ deploy/src/Microsoft.Azure.IIoT.Deployment/Infrastructure/MicrosoftGraphServiceClient.cs
@@ -16,16 +16,12 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
 
     class MicrosoftGraphServiceClient {
 
-        private readonly Guid _tenantGuid;
         private readonly GraphServiceClient _graphServiceClient;
 
         public MicrosoftGraphServiceClient(
-            Guid tenantGuid,
             ITokenProvider microsoftGraphTokenProvider,
             CancellationToken cancellationToken = default
         ) {
-            _tenantGuid = tenantGuid;
-
             var delegateAuthenticationProvider = new DelegateAuthenticationProvider(
                 async (httpRequestMessage) => {
                     var authenticationHeaderValue = await microsoftGraphTokenProvider
@@ -39,12 +35,9 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
         }
 
         public MicrosoftGraphServiceClient(
-            Guid tenantGuid,
             TokenCredentials microsoftGraphTokenCredentials,
             CancellationToken cancellationToken = default
         ) {
-            _tenantGuid = tenantGuid;
-
             var delegateAuthenticationProvider = new DelegateAuthenticationProvider(
                 async (httpRequestMessage) => {
                     await microsoftGraphTokenCredentials
@@ -77,417 +70,62 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
                 .GetAsync(cancellationToken);
         }
 
-        public async Task<ServicePrincipal> GetServicePrincipalByIdAsync(
-            Guid servicePrincipalId,
+        public async Task<Application> CreateApplicationAsync(
+            Application applicationDefinition,
             CancellationToken cancellationToken = default
         ) {
-            return await _graphServiceClient
-                .ServicePrincipals[servicePrincipalId.ToString()]
-                .Request()
-                .GetAsync(cancellationToken);
-        }
-
-        public async Task<Application> RegisterServiceApplicationAsync(
-            string servicesApplicationName,
-            DirectoryObject owner,
-            IEnumerable<string> tags = null,
-            CancellationToken cancellationToken = default
-        ) {
-            tags ??= new List<string>();
-
-            // Setup AppRoles for service application
-            var serviceApplicationAppRoles = new List<AppRole>();
-
-            var serviceApplicationApproverRoleIdGuid = Guid.NewGuid();
-            serviceApplicationAppRoles.Add(new AppRole {
-                DisplayName = "Approver",
-                Value = "Sign",
-                Description = "Approvers have the ability to issue certificates.",
-                AllowedMemberTypes = new List<string> { "User", "Application" },
-                Id = serviceApplicationApproverRoleIdGuid
-            });
-
-            var serviceApplicationWriterRoleIdGuid = Guid.NewGuid();
-            serviceApplicationAppRoles.Add(new AppRole {
-                DisplayName = "Writer",
-                Value = "Write",
-                Description = "Writers Have the ability to change entities.",
-                AllowedMemberTypes = new List<string> { "User", "Application" },
-                Id = serviceApplicationWriterRoleIdGuid
-            });
-
-            var serviceApplicationAdministratorRoleIdGuid = Guid.NewGuid();
-            serviceApplicationAppRoles.Add(new AppRole {
-                DisplayName = "Administrator",
-                Value = "Admin",
-                Description = "Admins can access advanced features.",
-                AllowedMemberTypes = new List<string> { "User", "Application" },
-                Id = serviceApplicationAdministratorRoleIdGuid
-            });
-
-            // Setup RequiredResourceAccess for service application
-
-            var keyVaultUserImpersonationRequiredResourceAccess = new RequiredResourceAccess {
-                ResourceAppId = AzureAppsConstants.AzureKeyVault.AppId,
-                ResourceAccess = new List<ResourceAccess> {
-                    new ResourceAccess {
-                        Id = AzureAppsConstants.AzureKeyVault.ResourceAccess["user_impersonation"],
-                        Type = "Scope"
-                    }
-                }
-            };
-
-            var microsoftGraphUserReadRequiredResourceAccess = new RequiredResourceAccess {
-                ResourceAppId = AzureAppsConstants.MicrosoftGraph.AppId,
-                ResourceAccess = new List<ResourceAccess> {
-                    new ResourceAccess {
-                        Id = AzureAppsConstants.MicrosoftGraph.ResourceAccess["User.Read"],
-                        Type = "Scope"
-                    }
-                }
-            };
-
-            var serviceApplicationRequiredResourceAccess = new List<RequiredResourceAccess>() {
-                keyVaultUserImpersonationRequiredResourceAccess,
-                microsoftGraphUserReadRequiredResourceAccess
-            };
-
-            // Add OAuth2Permissions
-            var oauth2Permissions = new List<PermissionScope> {
-                new PermissionScope {
-                    AdminConsentDescription = $"Allow the app to access {servicesApplicationName} on behalf of the signed-in user.",
-                    AdminConsentDisplayName = $"Access {servicesApplicationName}",
-                    Id = Guid.NewGuid(),
-                    IsEnabled = true,
-                    Type = "User",
-                    UserConsentDescription = $"Allow the application to access {servicesApplicationName} on your behalf.",
-                    UserConsentDisplayName = $"Access {servicesApplicationName}",
-                    Value = "user_impersonation"
-                }
-            };
-
-            var serviceApplicationApiApplication = new ApiApplication {
-                Oauth2PermissionScopes = oauth2Permissions
-            };
-
-            var serviceApplicationWebApplication = new WebApplication {
-                ImplicitGrantSettings = new ImplicitGrantSettings {
-                    EnableIdTokenIssuance = true
-                }
-            };
-
-            var serviceApplicationIdentifierUri = $"https://{_tenantGuid.ToString()}/{servicesApplicationName}";
-
-            var serviceApplicationDefinition = new Application {
-                DisplayName = servicesApplicationName,
-                IsFallbackPublicClient = false,
-                IdentifierUris = new List<string> { serviceApplicationIdentifierUri },
-                Tags = tags,
-                SignInAudience = "AzureADMyOrg",
-                AppRoles = serviceApplicationAppRoles,
-                RequiredResourceAccess = serviceApplicationRequiredResourceAccess,
-                Api = serviceApplicationApiApplication,
-                Web = serviceApplicationWebApplication,
-                PasswordCredentials = new List<PasswordCredential> { }
-            };
-
-            var serviceApplication = await _graphServiceClient
+            var application = await _graphServiceClient
                 .Applications
                 .Request()
-                .AddAsync(serviceApplicationDefinition, cancellationToken);
-
-            // Add Service Key PasswordCredential
-            var serviceKeyName = "Service Key";
-
-            var serviceApplicationServiceKeyPasswordCredentialDefinition = new PasswordCredential {
-                StartDateTime = DateTimeOffset.UtcNow,
-                EndDateTime = DateTimeOffset.UtcNow.AddYears(2),
-                CustomKeyIdentifier = ToBase64Bytes(serviceKeyName),
-                DisplayName = serviceKeyName
-            };
-
-            await _graphServiceClient
-                .Applications[serviceApplication.Id]
-                .AddPassword(serviceApplicationServiceKeyPasswordCredentialDefinition)
-                .Request()
-                .PostAsync(cancellationToken);
-
-            // We need to create ServicePrincipal for this application.
-            var serviceApplicationSP = await CreateServicePrincipalAsync(
-                serviceApplication,
-                tags,
-                cancellationToken
-            );
-
-            // Add app role assignment to owner as Approver, Writer and Administrator.
-            string PrincipalType;
-
-            if (owner is User) {
-                PrincipalType = "User";
-            } else if (owner is ServicePrincipal) {
-                PrincipalType = "ServicePrincipal";
-            } else if (owner is Group) {
-                PrincipalType = "Group";
-            } else {
-                throw new ArgumentException($"Owner is of unknown type: {owner.GetType()}");
-            }
-
-            var approverAppRoleAssignmentDefinition = new AppRoleAssignment {
-                //PrincipalDisplayName = "",
-                PrincipalType = PrincipalType,
-                PrincipalId = new Guid(owner.Id),
-                ResourceId = new Guid(serviceApplicationSP.Id),
-                ResourceDisplayName = "Approver",
-                Id = serviceApplicationApproverRoleIdGuid.ToString(),
-                AppRoleId = serviceApplicationApproverRoleIdGuid
-            };
-
-            var writerAppRoleAssignmentDefinition = new AppRoleAssignment {
-                //PrincipalDisplayName = "",
-                PrincipalType = PrincipalType,
-                PrincipalId = new Guid(owner.Id),
-                ResourceId = new Guid(serviceApplicationSP.Id),
-                ResourceDisplayName = "Writer",
-                Id = serviceApplicationWriterRoleIdGuid.ToString(),
-                AppRoleId = serviceApplicationWriterRoleIdGuid
-            };
-
-            var administratorAppRoleAssignmentDefinition = new AppRoleAssignment {
-                //PrincipalDisplayName = "",
-                PrincipalType = PrincipalType,
-                PrincipalId = new Guid(owner.Id),
-                ResourceId = new Guid(serviceApplicationSP.Id),
-                ResourceDisplayName = "Administrator",
-                Id = serviceApplicationAdministratorRoleIdGuid.ToString(),
-                AppRoleId = serviceApplicationAdministratorRoleIdGuid
-            };
-
-            await _graphServiceClient
-                .ServicePrincipals[serviceApplicationSP.Id]
-                .AppRoleAssignments
-                .Request()
                 .AddAsync(
-                    approverAppRoleAssignmentDefinition,
+                    applicationDefinition,
                     cancellationToken
                 );
 
-            await _graphServiceClient
-                .ServicePrincipals[serviceApplicationSP.Id]
-                .AppRoleAssignments
-                .Request()
-                .AddAsync(
-                    writerAppRoleAssignmentDefinition,
-                    cancellationToken
-                );
-
-            await _graphServiceClient
-                .ServicePrincipals[serviceApplicationSP.Id]
-                .AppRoleAssignments
-                .Request()
-                .AddAsync(
-                    administratorAppRoleAssignmentDefinition,
-                    cancellationToken
-                );
-
-            return serviceApplication;
+            return application;
         }
 
-        public async Task<Application> RegisterClientApplicationAsync(
-            Application serviceApplication,
-            string clientsApplicationName,
-            IEnumerable<string> tags = null,
+        public async Task<PasswordCredential> AddApplicationPasswordCredentialAsync(
+            Application application,
+            PasswordCredential passwordCredentialDefinition,
             CancellationToken cancellationToken = default
         ) {
-            tags ??= new List<string>();
-
-            // Extract id of Oauth2PermissionScope for user impersonation
-            var saApiOauth2PermissionScopeUserImpersonationList = serviceApplication
-                .Api
-                .Oauth2PermissionScopes
-                .Where(scope => "User" == scope.Type
-                    && "user_impersonation" == scope.Value
-                    && scope.IsEnabled.GetValueOrDefault(false)
-                )
-                .ToList();
-
-            if (saApiOauth2PermissionScopeUserImpersonationList.Count != 1
-                || !saApiOauth2PermissionScopeUserImpersonationList.First().Id.HasValue) {
-                throw new Exception("Service appplication does not expose Oauth2PermissionScope for user impersonation.");
-            }
-
-            var saApiOauth2PermissionScopeUserImpersonationId = 
-                saApiOauth2PermissionScopeUserImpersonationList.First().Id.Value;
-
-            var serviceApplicationUserImpersonationRequiredResourceAccess = new RequiredResourceAccess {
-                ResourceAppId = serviceApplication.AppId,  // service application
-                ResourceAccess = new List<ResourceAccess> {
-                    new ResourceAccess {
-                        Id = saApiOauth2PermissionScopeUserImpersonationId,  // "user_impersonation"
-                        Type = "Scope"
-                    }
-                }
-            };
-
-            var microsoftGraphUserReadRequiredResourceAccess = new RequiredResourceAccess {
-                ResourceAppId = AzureAppsConstants.MicrosoftGraph.AppId,
-                ResourceAccess = new List<ResourceAccess> {
-                    new ResourceAccess {
-                        Id = AzureAppsConstants.MicrosoftGraph.ResourceAccess["User.Read"],
-                        Type = "Scope"
-                    }
-                }
-            };
-
-            var clientApplicationRequiredResourceAccess = new List<RequiredResourceAccess>() {
-                serviceApplicationUserImpersonationRequiredResourceAccess,
-                microsoftGraphUserReadRequiredResourceAccess
-            };
-
-            var clientApplicationPublicClientApplication = new Microsoft.Graph.PublicClientApplication {
-                RedirectUris = new List<string> {
-                    "urn:ietf:wg:oauth:2.0:oob"
-                }
-            };
-
-            // Note: Oauth2AllowImplicitFlow will be enabled automatically since both
-            // EnableIdTokenIssuance and EnableAccessTokenIssuance are set to true.
-
-            var clientApplicationWebApplicatoin = new WebApplication {
-                //Oauth2AllowImplicitFlow = true,
-                ImplicitGrantSettings = new ImplicitGrantSettings {
-                    EnableIdTokenIssuance = true,
-                    EnableAccessTokenIssuance = true
-                }
-            };
-
-            var clientApplicationDefinition = new Application {
-                DisplayName = clientsApplicationName,
-                IsFallbackPublicClient = true,
-                Tags = tags,
-                SignInAudience = "AzureADMyOrg",
-                RequiredResourceAccess = clientApplicationRequiredResourceAccess,
-                PublicClient = clientApplicationPublicClientApplication,
-                Web = clientApplicationWebApplicatoin,
-                PasswordCredentials = new List<PasswordCredential> { }
-            };
-
-            var clientApplication = await _graphServiceClient
-                .Applications
-                .Request()
-                .AddAsync(clientApplicationDefinition, cancellationToken);
-
-            // Add Client Key PasswordCredential
-            var clientKeyName = "Client Key";
-
-            var clientApplicationClientKeyPasswordCredentialDefinition = new PasswordCredential {
-                StartDateTime = DateTimeOffset.UtcNow,
-                EndDateTime = DateTimeOffset.UtcNow.AddYears(2),
-                CustomKeyIdentifier = ToBase64Bytes(clientKeyName),
-                DisplayName = clientKeyName
-            };
-
-            await _graphServiceClient
-                .Applications[clientApplication.Id]
-                .AddPassword(clientApplicationClientKeyPasswordCredentialDefinition)
+            var passwordCredential = await _graphServiceClient
+                .Applications[application.Id]
+                .AddPassword(passwordCredentialDefinition)
                 .Request()
                 .PostAsync(cancellationToken);
 
-            // We need to create ServicePrincipal for this application.
-            await CreateServicePrincipalAsync(
-                clientApplication,
-                tags,
-                cancellationToken
-            );
-
-            return clientApplication;
+            return passwordCredential;
         }
 
-        public async Task<Tuple<Application, string>> RegisterAKSApplicationAsync(
-            string aksApplicationName,
-            IEnumerable<string> tags = null,
+        /// <summary>
+        /// Add PasswordCredential to an Application that is valid for
+        /// 2 years starting from now.
+        /// </summary>
+        /// <param name="application">Application to which PasswordCredential will be added.</param>
+        /// <param name="keyName">Display name of PasswordCredential.</param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        public async Task<PasswordCredential> AddApplication2YPasswordCredentialAsync(
+            Application application,
+            string keyName,
             CancellationToken cancellationToken = default
         ) {
-            tags ??= new List<string>();
-
-            // Add OAuth2Permissions for user impersonation
-            var aksOauth2Permissions = new List<PermissionScope> {
-                new PermissionScope {
-                    AdminConsentDescription = $"Allow the app to access {aksApplicationName} on behalf of the signed-in user.",
-                    AdminConsentDisplayName = $"Access {aksApplicationName}",
-                    Id = Guid.NewGuid(),
-                    IsEnabled = true,
-                    Type = "User",
-                    UserConsentDescription = $"Allow the application to access {aksApplicationName} on your behalf.",
-                    UserConsentDisplayName = $"Access {aksApplicationName}",
-                    Value = "user_impersonation"
-                }
-            };
-
-            var aksApplicationApiApplication = new ApiApplication {
-                Oauth2PermissionScopes = aksOauth2Permissions
-            };
-
-            var aksApplicationWebApplication = new WebApplication {
-                ImplicitGrantSettings = new ImplicitGrantSettings {
-                    EnableIdTokenIssuance = true
-                }
-            };
-
-            var aksApplicationIdentifierUri = $"https://{_tenantGuid.ToString()}/{aksApplicationName}";
-
-            var aksApplicationDefinition = new Application {
-                DisplayName = aksApplicationName,
-                IsFallbackPublicClient = false,
-                IdentifierUris = new List<string> { aksApplicationIdentifierUri },
-                Tags = tags,
-                SignInAudience = "AzureADMyOrg",
-                AppRoles = new List<AppRole>(),
-                RequiredResourceAccess = new List<RequiredResourceAccess>(),
-                Api = aksApplicationApiApplication,
-                Web = aksApplicationWebApplication,
-                PasswordCredentials = new List<PasswordCredential> { }
-            };
-
-            var aksApplication = await _graphServiceClient
-                .Applications
-                .Request()
-                .AddAsync(aksApplicationDefinition, cancellationToken);
-
-            // Add RBAC Key PasswordCredential
-            var rbacKeyName = "rbac";
-
-            var aksApplicationRBACPasswordCredentialDefinition = new PasswordCredential {
+            var passwordCredentialDefinition = new PasswordCredential {
                 StartDateTime = DateTimeOffset.UtcNow,
                 EndDateTime = DateTimeOffset.UtcNow.AddYears(2),
-                CustomKeyIdentifier = ToBase64Bytes(rbacKeyName),
-                DisplayName = rbacKeyName
+                CustomKeyIdentifier = ToBase64Bytes(keyName),
+                DisplayName = keyName
             };
 
-            var aksApplicationRBACPasswordCredential = await _graphServiceClient
-                .Applications[aksApplication.Id]
-                .AddPassword(aksApplicationRBACPasswordCredentialDefinition)
-                .Request()
-                .PostAsync(cancellationToken);
-
-            if (string.IsNullOrEmpty(aksApplicationRBACPasswordCredential.SecretText)) {
-                throw new Exception($"Failed to retrieve password credentials for AKS Application: {rbacKeyName}");
-            }
-
-            var aksApplicationRBACPasswordCredentialSecret = aksApplicationRBACPasswordCredential.SecretText;
-
-            // We need to create ServicePrincipal for this application.
-            await CreateServicePrincipalAsync(
-                aksApplication,
-                tags,
+            var passwordCredential = await AddApplicationPasswordCredentialAsync(
+                application,
+                passwordCredentialDefinition,
                 cancellationToken
-            );
+             );
 
-            var result = new Tuple<Application, string>(
-                aksApplication,
-                aksApplicationRBACPasswordCredentialSecret
-            );
-
-            return result;
+            return passwordCredential;
         }
 
         /// <summary>
@@ -568,7 +206,7 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
         /// <param name="redirectUris"></param>
         /// <param name="cancellationToken"></param>
         /// <returns>Updated application.</returns>
-        public async Task<Application> UpdateRedirectUrisAsync(
+        public async Task<Application> UpdateApplicationRedirectUrisAsync(
             Application application,
             IEnumerable<string> redirectUris,
             CancellationToken cancellationToken = default
@@ -600,6 +238,39 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
                 Log.Error(ex, $"Failed to update RedirectUris of application: {application.DisplayName}");
                 throw;
             }
+        }
+
+        /// <summary>
+        /// Get ServicePrincipal by its Id.
+        /// </summary>
+        /// <param name="servicePrincipalId">Guid representing servicePrincipal Id.</param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        public async Task<ServicePrincipal> GetServicePrincipalAsync(
+            Guid servicePrincipalId,
+            CancellationToken cancellationToken = default
+        ) {
+            return await _graphServiceClient
+                .ServicePrincipals[servicePrincipalId.ToString()]
+                .Request()
+                .GetAsync(cancellationToken);
+        }
+
+        public async Task<AppRoleAssignment> AddServicePrincipalAppRoleAssignmentAsync(
+            ServicePrincipal servicePrincipal,
+            AppRoleAssignment appRoleAssignmentDefinition,
+            CancellationToken cancellationToken = default
+        ) {
+            var appRoleAssignment = await _graphServiceClient
+                .ServicePrincipals[servicePrincipal.Id]
+                .AppRoleAssignments
+                .Request()
+                .AddAsync(
+                    appRoleAssignmentDefinition,
+                    cancellationToken
+                );
+
+            return appRoleAssignment;
         }
 
         /// <summary>
@@ -637,7 +308,7 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
         /// <param name="application"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
-        public async Task<ServicePrincipal> GetServicePrincipalAsync(
+        public async Task<ServicePrincipal> GetApplicationServicePrincipalAsync(
             Application application,
             CancellationToken cancellationToken = default
         ) {
@@ -647,7 +318,7 @@ namespace Microsoft.Azure.IIoT.Deployment.Infrastructure {
             );
         }
 
-        public async Task<ServicePrincipal> CreateServicePrincipalAsync(
+        public async Task<ServicePrincipal> CreateApplicationServicePrincipalAsync(
             Application application,
             IEnumerable<string> tags = null,
             CancellationToken cancellationToken = default
